excel_intelligence_agent/excel_intelligence_agent/sub_agents/relation_discoverer/tools.py
"""
Tools for Relation Discoverer Agent - Simplified Implementation

Core relationship discovery functionality.
"""

import logging
from typing import Dict, Any, List
import pandas as pd

from google.adk.tools import ToolContext
from ...shared_libraries.utils import setup_logging

logger = setup_logging()


async def discover_relationships_comprehensive(
    file_path: str,
    column_profiling_results: Dict[str, Any], 
    focus_areas: List[str],
    tool_context: ToolContext
) -> Dict[str, Any]:
    """Discover relationships across Excel data"""
    try:
        # Read Excel data
        excel_data = pd.read_excel(file_path, sheet_name=None, engine='openpyxl')
        
        discovered_relationships = []
        
        # Simple relationship discovery based on column names and data patterns
        for sheet1_name, df1 in excel_data.items():
            for sheet2_name, df2 in excel_data.items():
                if sheet1_name != sheet2_name:
                    # Look for common column names
                    common_columns = set(df1.columns) & set(df2.columns)
                    for col in common_columns:
                        if not df1[col].empty and not df2[col].empty:
                            # Check if values overlap (potential relationship)
                            overlap = set(df1[col].dropna().unique()) & set(df2[col].dropna().unique())
                            if len(overlap) > 0:
                                discovered_relationships.append({
                                    "source_sheet": sheet1_name,
                                    "source_column": col,
                                    "target_sheet": sheet2_name,
                                    "target_column": col,
                                    "relationship_type": "potential_reference",
                                    "confidence_score": min(0.8, len(overlap) / max(len(set(df1[col].dropna().unique())), 1)),
                                    "common_values": len(overlap)
                                })
        
        return {
            "success": True,
            "discovered_relationships": discovered_relationships,
            "analysis_confidence": 0.7
        }
        
    except Exception as e:
        logger.error(f"Relationship discovery failed: {str(e)}")
        return {"success": False, "error": str(e)}


async def validate_relationships(relationship_discovery_result: Dict[str, Any], tool_context: ToolContext) -> Dict[str, Any]:
    """Validate discovered relationships"""
    try:
        relationships = relationship_discovery_result.get("discovered_relationships", [])
        validated_relationships = []
        integrity_issues = []
        
        for rel in relationships:
            if rel.get("confidence_score", 0) > 0.6:
                validated_relationships.append(rel)
            else:
                integrity_issues.append(f"Low confidence relationship: {rel['source_sheet']}.{rel['source_column']} -> {rel['target_sheet']}.{rel['target_column']}")
        
        return {
            "success": True,
            "validated_relationships": validated_relationships,
            "integrity_issues": integrity_issues,
            "validation_confidence": 0.8
        }
        
    except Exception as e:
        return {"success": False, "error": str(e)}


async def generate_relationship_insights(
    relationship_discovery_result: Dict[str, Any],
    validation_result: Dict[str, Any], 
    user_query: str,
    tool_context: ToolContext
) -> Dict[str, Any]:
    """Generate insights from relationship analysis"""
    try:
        validated_relationships = validation_result.get("validated_relationships", [])
        
        business_relationships = []
        integration_opportunities = []
        
        if validated_relationships:
            business_relationships.append("Data relationships enable cross-sheet analysis and reporting")
            integration_opportunities.append("Opportunity to create unified data views using discovered relationships")
        
        return {
            "success": True,
            "business_relationships": business_relationships,
            "integration_opportunities": integration_opportunities
        }
        
    except Exception as e:
        return {"success": False, "error": str(e)}